backend/requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.35,<1.0
sqlalchemy>=2.0,<3.0
alembic>=1.14,<2.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
python-jose[cryptography]>=3.3,<4.0
argon2-cffi>=23.1,<26.0
bcrypt>=4.1,<5.0
email-validator>=2.2,<3.0
python-multipart>=0.0.9,<1.0
pandas>=2.2,<3.0
numpy>=1.26,<3.0
scikit-learn>=1.5,<2.0
openpyxl>=3.1,<4.0